+++ CHANGELOG.md
@@ -10,6 +10,14 @@ This is the version currently published on master.
   + GET Single
   + GET List
 
+### [Under Development] - Changed
+
++ `file_utils` now uses a interface to better describe the expected `json` files
+
+### [Under Development] - Fixed
+
+### [Under Development] - Deprecated
+
 ## 🆕 [v0.0.1]
 
 The first and simplest version of the api.